Added single diffusion and edge weights to TieDie
Added single diffusion option to TieDie, which ignores the down_heats and proceeds with only the up_heats.

When using pagerank diffusion only, we also added new feature to set edge weights in the network according to the supplied edge scores. Edge scores are positive, in our application they are -log10 of p-values from an experiment, and they control the edge weights via the lower_w, upper_w. and max_penalty flags. Edge weights are between 0 and 1, and are equal to 1 by default, but can be lowered all the way to zero by using max_penalty = 1, with max_penalty = 0 resulting in no penalty. Edge score below lower_w are applied the max_penalty, while scores above upper_w receive no penalty. In this way, we can easily use the Makefile to get different flow conditions in the network.

Finally, set permutations to optional to save on computation time, and added the saved_diffusion flag to save the diffused heats to the working folder to make it easy to test different size and alpha paramaters. Set to to False by default.